Implemented MVC Controller for Restaurants
The previous controller that was written was a REST Controller thus it was refactored to RestaurantRESTController. A new controller was created with its original name in order to use MVC to return a view with a web page containing a table with the contents of all rows in the database.

Thymeleaf was added to the Maven dependencies and used to create a HTML template to display the table with all restaurants. Bootstrap was included to bootstrap the css style for the table and page.